requirements.txt
asn1crypto==0.24.0
certifi==2019.3.9
cffi==1.12.3
cryptography==2.6.1
future==0.17.1
geographiclib==1.49
geopy==1.19.0
importlib-resources==1.0.2
numpy==1.16.3
pycparser==2.19
python-dotenv==0.10.2
python-telegram-bot==11.1.0
pytz==2019.1
six==1.12.0
timezonefinder==4.0.2
